pc_stats/requirements-host.txt
psutil
pyserial
pynvml
python-dotenv
